// frontend/src/client/types.gen.ts
export type ItemCreate = {
    title: string;
    description?: (string | null);
};